+++ src/main/java/com/grupo06/Resourses/Profesor.java
@@ -4,8 +4,6 @@
  */
 package com.grupo06.Resourses;
 
-import java.math.BigDecimal;
-
 /**
  *
  * @author santi
@@ -14,16 +12,16 @@ public class Profesor implements java.io.Serializable{
     private Integer id;
     private String nombre;
     private String apellidos;
-    private BigDecimal sueldo;
+    private int sueldo;
 
-    public Profesor(String nombre, String apellidos, BigDecimal sueldo) {
+    public Profesor(String nombre, String apellidos, int sueldo) {
         this.nombre = nombre;
         this.apellidos = apellidos;
         this.sueldo = sueldo;
         this.id = null;
     }
 
-    public Profesor(Integer id, String nombre, String apellidos, BigDecimal sueldo) {
+    public Profesor(Integer id, String nombre, String apellidos, int sueldo) {
         this.id = id;
         this.nombre = nombre;
         this.apellidos = apellidos;
@@ -54,11 +52,11 @@ public class Profesor implements java.io.Serializable{
         this.apellidos = apellidos;
     }
 
-    public BigDecimal getSueldo() {
+    public int getSueldo() {
         return sueldo;
     }
 
-    public void setSueldo(BigDecimal sueldo) {
+    public void setSueldo(int sueldo) {
         this.sueldo = sueldo;
     }
